requirements.txt
pyTelegramBotAPI==4.10.0
wikipedia==1.4.0
python-decouple==3.8
